+++ db/schema.rb
@@ -2,16 +2,15 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_12_14_132751) do
-
+ActiveRecord::Schema[7.0].define(version: 2022_12_14_132751) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
@@ -35,8 +34,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "author_type"
     t.integer "author_id"
     t.text "body"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "namespace"
     t.index ["author_type", "author_id"], name: "index_active_admin_comments_on_author_type_and_author_id"
     t.index ["namespace"], name: "index_active_admin_comments_on_namespace"
@@ -49,12 +48,12 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "user_id"
     t.boolean "completed", default: false, null: false
     t.float "progress", default: 0.0, null: false
-    t.datetime "completed_at"
+    t.datetime "completed_at", precision: nil
     t.boolean "successful"
     t.text "error_message"
     t.jsonb "results", default: {}, null: false
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.enum "state", default: "scheduled", null: false, enum_type: "background_jobs_state_type"
     t.index ["completed"], name: "index_background_jobs_on_completed"
     t.index ["legacy_id"], name: "index_background_jobs_on_legacy_id"
@@ -67,12 +66,12 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "session_id"
     t.integer "patient_id"
     t.string "images"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "case_type"
     t.integer "state", default: 0
     t.integer "flag", default: 0
-    t.datetime "exported_at"
+    t.datetime "exported_at", precision: nil
     t.boolean "no_export", default: false
     t.string "comment"
     t.integer "assigned_reader_id"
@@ -89,8 +88,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
   create_table "centers", id: :serial, force: :cascade do |t|
     t.string "name"
     t.integer "study_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "code"
     t.string "domino_unid"
     t.index ["study_id", "code"], name: "index_centers_on_study_id_and_code", unique: true
@@ -103,8 +102,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "configurable_type", null: false
     t.uuid "configurable_id", null: false
     t.enum "schema_spec", null: false, comment: "Specify the configuration schema for the given `configuration_type`.\n", enum_type: "configuration_schema_specs"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["configurable_id"], name: "index_configurations_on_configurable_id"
     t.index ["previous_configuration_id"], name: "index_configurations_on_previous_configuration_id"
   end
@@ -113,8 +112,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "name", null: false
     t.string "email_type", null: false
     t.text "template", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "form_answer_resources", force: :cascade do |t|
@@ -137,18 +136,18 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "annotated_images_signature", comment: "RSA Signature via private part of `public_key`.\n"
     t.boolean "is_test_data", default: false, null: false
     t.boolean "is_obsolete", default: false, null: false
-    t.datetime "signed_at"
-    t.datetime "submitted_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "signed_at", precision: nil
+    t.datetime "submitted_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "user_id"
     t.integer "study_id"
     t.integer "form_session_id"
     t.integer "form_display_type_id"
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.integer "sequence_number", default: 0, null: false
     t.integer "blocking_user_id"
-    t.datetime "blocked_at"
+    t.datetime "blocked_at", precision: nil
     t.index ["configuration_id"], name: "index_form_answers_on_configuration_id"
     t.index ["form_definition_id"], name: "index_form_answers_on_form_definition_id"
     t.index ["public_key_id"], name: "index_form_answers_on_public_key_id"
@@ -158,25 +157,25 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "name", null: false
     t.text "description", null: false
     t.uuid "locked_configuration_id"
-    t.datetime "locked_at"
+    t.datetime "locked_at", precision: nil
     t.uuid "current_configuration_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "form_sessions", force: :cascade do |t|
     t.string "name", null: false
     t.string "description"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["name"], name: "index_form_sessions_on_name"
   end
 
   create_table "forms", id: :serial, force: :cascade do |t|
     t.string "name"
     t.text "description"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.integer "session_id"
     t.integer "state", default: 0
     t.string "locked_version"
@@ -186,7 +185,7 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
   create_table "historic_report_cache_entries", id: :serial, force: :cascade do |t|
     t.integer "historic_report_query_id", null: false
     t.integer "study_id", null: false
-    t.datetime "date", null: false
+    t.datetime "date", precision: nil, null: false
     t.integer "version_id"
     t.index ["date"], name: "index_historic_report_cache_entries_on_date"
     t.index ["historic_report_query_id"], name: "index_historic_report_cache_entries_on_historic_report_query_id"
@@ -204,15 +203,15 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
   create_table "historic_report_queries", id: :serial, force: :cascade do |t|
     t.string "resource_type"
     t.string "group_by"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "image_series", id: :serial, force: :cascade do |t|
     t.string "name"
     t.integer "visit_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.integer "patient_id"
     t.date "imaging_date"
     t.string "domino_unid"
@@ -221,6 +220,7 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "comment"
     t.jsonb "properties", default: {}, null: false
     t.string "properties_version"
+    t.jsonb "cache", default: {}, null: false
     t.index ["patient_id", "series_number"], name: "index_image_series_on_patient_id_and_series_number"
     t.index ["patient_id"], name: "index_image_series_on_patient_id"
     t.index ["series_number"], name: "index_image_series_on_series_number"
@@ -229,10 +229,11 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
 
   create_table "images", id: :serial, force: :cascade do |t|
     t.integer "image_series_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "mimetype"
     t.string "sha256sum"
+    t.jsonb "cache", default: {}, null: false
     t.index ["image_series_id"], name: "index_images_on_image_series_id"
   end
 
@@ -258,8 +259,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.jsonb "filters", default: [], null: false
     t.boolean "only_authorized_recipients", default: true, null: false
     t.boolean "is_enabled", default: false, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "maximum_email_throttling_delay"
     t.jsonb "triggering_actions", default: [], null: false
     t.integer "email_template_id", null: false
@@ -272,10 +273,10 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "resource_id"
     t.integer "version_id"
     t.integer "user_id", null: false
-    t.datetime "email_sent_at"
-    t.datetime "marked_seen_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "email_sent_at", precision: nil
+    t.datetime "marked_seen_at", precision: nil
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "triggering_action", null: false
     t.index ["resource_type", "resource_id"], name: "index_notifications_on_resource_type_and_resource_id"
     t.index ["user_id"], name: "index_notifications_on_user_id"
@@ -285,12 +286,13 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
   create_table "patients", id: :serial, force: :cascade do |t|
     t.string "subject_id"
     t.string "images_folder"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.integer "center_id"
     t.string "domino_unid"
     t.jsonb "data", default: {}, null: false
     t.jsonb "export_history", default: [], null: false
+    t.jsonb "cache", default: {}, null: false
     t.index ["center_id"], name: "index_patients_on_center_id"
   end
 
@@ -298,8 +300,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "role_id", null: false
     t.string "activity", null: false
     t.string "subject", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["activity"], name: "index_permissions_on_activity"
     t.index ["role_id"], name: "index_permissions_on_role_id"
     t.index ["subject"], name: "index_permissions_on_subject"
@@ -309,9 +311,9 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "user_id", null: false
     t.text "public_key", null: false
     t.boolean "active", null: false
-    t.datetime "deactivated_at"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "deactivated_at", precision: nil
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.index ["active"], name: "index_public_keys_on_active"
     t.index ["user_id"], name: "index_public_keys_on_user_id"
   end
@@ -327,29 +329,29 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "name", null: false
     t.integer "image_series_id"
     t.integer "tqc_state"
-    t.datetime "tqc_date"
+    t.datetime "tqc_date", precision: nil
     t.string "tqc_version"
     t.jsonb "tqc_results"
     t.integer "tqc_user_id"
     t.text "tqc_comment"
     t.string "domino_unid"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["image_series_id"], name: "index_required_series_on_image_series_id"
     t.index ["visit_id", "name"], name: "index_required_series_on_visit_id_and_name", unique: true
   end
 
   create_table "roles", id: :serial, force: :cascade do |t|
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "title", null: false
   end
 
   create_table "sessions", id: :serial, force: :cascade do |t|
     t.string "name"
     t.integer "study_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.integer "state", default: 0
     t.string "locked_version"
     t.index ["study_id"], name: "index_sessions_on_study_id"
@@ -357,8 +359,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
 
   create_table "studies", id: :serial, force: :cascade do |t|
     t.string "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "locked_version"
     t.string "domino_db_url"
     t.string "notes_links_base_uri"
@@ -373,7 +375,7 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "tagger_type"
     t.integer "tagger_id"
     t.string "context", limit: 128
-    t.datetime "created_at"
+    t.datetime "created_at", precision: nil
     t.index ["tag_id", "taggable_id", "taggable_type", "context", "tagger_id", "tagger_type"], name: "taggings_idx", unique: true
     t.index ["taggable_id", "taggable_type", "context"], name: "index_taggings_on_taggable_id_and_taggable_type_and_context"
   end
@@ -389,8 +391,8 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "role_id", null: false
     t.string "scope_object_type"
     t.integer "scope_object_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["role_id"], name: "index_user_roles_on_role_id"
     t.index ["scope_object_type", "scope_object_id"], name: "index_user_roles_on_scope_object_type_and_scope_object_id"
     t.index ["user_id"], name: "index_user_roles_on_user_id"
@@ -400,28 +402,28 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
-    t.datetime "reset_password_sent_at"
-    t.datetime "remember_created_at"
+    t.datetime "reset_password_sent_at", precision: nil
+    t.datetime "remember_created_at", precision: nil
     t.integer "sign_in_count", default: 0
-    t.datetime "current_sign_in_at"
-    t.datetime "last_sign_in_at"
+    t.datetime "current_sign_in_at", precision: nil
+    t.datetime "last_sign_in_at", precision: nil
     t.string "current_sign_in_ip"
     t.string "last_sign_in_ip"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "name"
     t.text "public_key"
     t.text "private_key"
     t.string "username"
-    t.datetime "password_changed_at"
+    t.datetime "password_changed_at", precision: nil
     t.string "authentication_token"
     t.integer "failed_attempts", default: 0
     t.string "unlock_token"
-    t.datetime "locked_at"
+    t.datetime "locked_at", precision: nil
     t.boolean "is_root_user", default: false, null: false
     t.string "confirmation_token"
-    t.datetime "confirmed_at"
-    t.datetime "confirmation_sent_at"
+    t.datetime "confirmed_at", precision: nil
+    t.datetime "confirmation_sent_at", precision: nil
     t.string "unconfirmed_email"
     t.integer "email_throttling_delay"
     t.jsonb "dashboard_configuration"
@@ -441,7 +443,7 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.string "item_id", null: false
     t.string "event", null: false
     t.string "whodunnit"
-    t.datetime "created_at"
+    t.datetime "created_at", precision: nil
     t.jsonb "object"
     t.jsonb "object_changes"
     t.integer "study_id"
@@ -463,11 +465,11 @@ ActiveRecord::Schema.define(version: 2022_12_14_132751) do
     t.integer "visit_number"
     t.string "visit_type"
     t.integer "patient_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.string "domino_unid"
     t.string "description"
-    t.datetime "mqc_date"
+    t.datetime "mqc_date", precision: nil
     t.integer "mqc_user_id"
     t.integer "state", default: 0
     t.integer "mqc_state", default: 0